Move some date utils to utils
