+++ packages/data-mate/src/core/index.ts
@@ -1,5 +1,4 @@
 export * from './config';
-export * from './date-utils';
 export * from './interfaces';
 export * from './ReadableData';
 export * from './utils';
